refactor: a settings object so future settings management will be easier
